break out credentials into its own module
