+++ dbt/adapters/bigquery/dataproc/batch.py
@@ -10,7 +10,7 @@ from google.cloud.dataproc_v1 import (
 )
 from google.protobuf.json_format import ParseDict
 
-from dbt.adapters.bigquery.connections import DataprocBatchConfig
+from dbt.adapters.bigquery.credentials import DataprocBatchConfig
 
 _BATCH_RUNNING_STATES = [Batch.State.PENDING, Batch.State.RUNNING]
 DEFAULT_JAR_FILE_URI = "gs://spark-lib/bigquery/spark-bigquery-with-dependencies_2.13-0.34.0.jar"
